Update integration for langchain to work.
