+++ integrations/langchain-memgraph/langchain_memgraph/__init__.py
@@ -2,7 +2,7 @@ from importlib import metadata
 
 from langchain_memgraph.chains.graph_qa import MemgraphQAChain
 from langchain_memgraph.document_loaders import MemgraphLoader
-from langchain_memgraph.graphs.memgraph import Memgraph
+from langchain_memgraph.graphs.memgraph import MemgraphLangchain
 from langchain_memgraph.retrievers import MemgraphRetriever
 from langchain_memgraph.toolkits import MemgraphToolkit
 from langchain_memgraph.tools import QueryMemgraphTool
@@ -17,7 +17,7 @@ del metadata  # optional, avoids polluting the results of dir(__package__)
 __all__ = [
     "MemgraphLoader",
     "MemgraphQAChain",
-    "Memgraph",
+    "MemgraphLangchain",
     "MemgraphRetriever",
     "MemgraphToolkit",
     "QueryMemgraphTool",
