+++ integrations/langchain-memgraph/pyproject.toml
@@ -7,7 +7,6 @@ readme = "README.md"
 license = { text = "MIT" }
 requires-python = ">=3.13"
 dependencies = [
-    
     "langchain-core>=0.3.15",
     "neo4j>=5.28.1",
 ]
